Fixed warning.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@32058 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/servers/app/FontManager.cpp
@@ -511,7 +511,7 @@ FontManager::_PrecacheFontFile(const ServerFont* font)
 	while (true) {
 		// We just want the file in the kernel file cache...
 		ssize_t read = file.Read(buffer, bufferSize);
-		if (read < bufferSize)
+		if (read < (ssize_t) bufferSize)
 			break;
 	}
